# frontend/api-docs.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Pharmacy"},"host":"localhost:8379","basePath":"/","tags":[{"name":"account-controller","description":"Account Controller"},{"name":"center-controller","description":"Center Controller"},{"name":"drug-controller","description":"Drug Controller"},{"name":"pharmacy-controller","description":"Pharmacy Controller"},{"name":"stock-controller","description":"Stock Controller"},{"name":"supplier-controller","description":"Supplier Controller"}],"paths":{"/account":{"get":{"tags":["account-controller"],"summary":"getAllAccounts","operationId":"getAllAccountsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Account"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/account/create":{"post":{"tags":["account-controller"],"summary":"createAccount","operationId":"createAccountUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"accountDto","description":"accountDto","required":true,"schema":{"$ref":"#/definitions/Account"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Account"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/account/update":{"put":{"tags":["account-controller"],"summary":"updateAccount","operationId":"updateAccountUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"accountDto","description":"accountDto","required":true,"schema":{"$ref":"#/definitions/Account"}}],"responses":{"204":{"description":"No Content"}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/account/{email}":{"get":{"tags":["account-controller"],"summary":"getByEmail","operationId":"getByEmailUsingGET","produces":["application/json"],"parameters":[{"name":"email","in":"path","description":"email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Account"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/account/{email}/isUnique":{"get":{"tags":["account-controller"],"summary":"isUnique","operationId":"isUniqueUsingGET","produces":["application/json"],"parameters":[{"name":"email","in":"path","description":"email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/account/{email}/roles":{"get":{"tags":["account-controller"],"summary":"getAccountRoles","operationId":"getAccountRolesUsingGET","produces":["application/json"],"parameters":[{"name":"email","in":"path","description":"email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"string"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center":{"get":{"tags":["center-controller"],"summary":"getAllCenters","operationId":"getAllCentersUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Center"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center/create":{"post":{"tags":["center-controller"],"summary":"createCenter","operationId":"createCenterUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"centerDto","description":"centerDto","required":true,"schema":{"$ref":"#/definitions/Center"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Center"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center/getByName":{"get":{"tags":["center-controller"],"summary":"getCentersByName","operationId":"getCentersByNameUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Center"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center/pharmacy/{id}":{"get":{"tags":["center-controller"],"summary":"getCentersByPharmacyId","operationId":"getCentersByPharmacyIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Center"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center/update":{"put":{"tags":["center-controller"],"summary":"updateCenter","operationId":"updateCenterUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"centerDto","description":"centerDto","required":true,"schema":{"$ref":"#/definitions/Center"}}],"responses":{"204":{"description":"No Content"}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/center/{id}":{"get":{"tags":["center-controller"],"summary":"getCenterById","operationId":"getCenterByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Center"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/drug":{"get":{"tags":["drug-controller"],"summary":"getAllDrugs","operationId":"getAllDrugsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Drug"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/drug/create":{"post":{"tags":["drug-controller"],"summary":"createDrug","operationId":"createDrugUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"drugDto","description":"drugDto","required":true,"schema":{"$ref":"#/definitions/Drug"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Drug"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/drug/getDrugs":{"get":{"tags":["drug-controller"],"summary":"getDrugsByIds","operationId":"getDrugsByIdsUsingGET","produces":["application/json"],"parameters":[{"name":"ids","in":"query","description":"ids","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Drug"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/drug/update":{"put":{"tags":["drug-controller"],"summary":"updateDrug","operationId":"updateDrugUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"drugDto","description":"drugDto","required":true,"schema":{"$ref":"#/definitions/Drug"}}],"responses":{"204":{"description":"No Content"}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/pharmacy":{"get":{"tags":["pharmacy-controller"],"summary":"getAllPharmacies","operationId":"getAllPharmaciesUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Pharmacy"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/pharmacy/create":{"post":{"tags":["pharmacy-controller"],"summary":"createPharmacy","operationId":"createPharmacyUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"pharmacyDto","description":"pharmacyDto","required":true,"schema":{"$ref":"#/definitions/Pharmacy"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Pharmacy"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/pharmacy/filterByName":{"get":{"tags":["pharmacy-controller"],"summary":"findByNameFilter","operationId":"findByNameFilterUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pharmacy"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/pharmacy/findByName":{"get":{"tags":["pharmacy-controller"],"summary":"findByName","operationId":"findByNameUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pharmacy"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/pharmacy/{id}":{"get":{"tags":["pharmacy-controller"],"summary":"findPharmacyById","operationId":"findPharmacyByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Pharmacy"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/stock":{"get":{"tags":["stock-controller"],"summary":"getAllStocks","operationId":"getAllStocksUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Stock"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/stock/center/{id}":{"get":{"tags":["stock-controller"],"summary":"getStocksByCenterId","operationId":"getStocksByCenterIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Stock"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/stock/create":{"post":{"tags":["stock-controller"],"summary":"createStock","operationId":"createStockUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"stockDto","description":"stockDto","required":true,"schema":{"$ref":"#/definitions/Stock"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Stock"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/stock/getByName":{"get":{"tags":["stock-controller"],"summary":"getByName","operationId":"getByNameUsingGET","produces":["application/json"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Stock"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/supplier":{"get":{"tags":["supplier-controller"],"summary":"getAllSuppliers","operationId":"getAllSuppliersUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Supplier"}}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}},"/supplier/create":{"post":{"tags":["supplier-controller"],"summary":"createSupplier","operationId":"createSupplierUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"supplierDto","description":"supplierDto","required":true,"schema":{"$ref":"#/definitions/Supplier"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Supplier"}},"401":{"description":"Unauthorized","schema":{"type":"string"}},"406":{"description":"Not Acceptable","schema":{"type":"string"}},"500":{"description":"500 message","schema":{"type":"string"}}},"security":[{"oauth2schema":["read","trust","write"]}],"deprecated":false}}},"securityDefinitions":{"oauth2schema":{"type":"oauth2","tokenUrl":"/oauth/token","flow":"password","scopes":{"read":"read all","trust":"trust all","write":"access all"}}},"definitions":{"Account":{"type":"object","properties":{"country":{"type":"string"},"county":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"type":"string"},"profilePicture":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}},"title":"Account"},"Center":{"type":"object","properties":{"city":{"type":"string"},"county":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"pharmacyId":{"type":"integer","format":"int64"},"phone":{"type":"string"},"postalCode":{"type":"string"},"street":{"type":"string"}},"title":"Center"},"Drug":{"type":"object","properties":{"description":{"type":"string"},"enabled":{"type":"boolean"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"integer","format":"int32"},"strength":{"type":"string"},"valability":{"type":"string","format":"date"}},"title":"Drug"},"Pharmacy":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"pharmacyType":{"type":"string","enum":["COMMUNITY","NATURIST"]},"supplierIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"title":"Pharmacy"},"Stock":{"type":"object","properties":{"centerId":{"type":"integer","format":"int64"},"drugId":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"nameOfStock":{"type":"string"},"totalQuantity":{"type":"integer","format":"int32"}},"title":"Stock"},"Supplier":{"type":"object","properties":{"city":{"type":"string"},"country":{"type":"string"},"county":{"type":"string"},"drugIds":{"type":"array","items":{"type":"integer","format":"int64"}},"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"phone":{"type":"string"},"street":{"type":"string"}},"title":"Supplier"}}}